+++ CHANGELOG.md
@@ -28,6 +28,8 @@
 - Multiple subjects are now supported for case expressions in JavaScript.
 - Fixed a bug where matching using a Bool or Nil literal as the subject for a
   case expression would produce invalid code when compiling to JavaScript.
+- Unsupported feature error messages now include file path and line numbers for
+  debugging.
 
 
 ## v0.16.1 - 2021-06-21